requirements.txt
Flask==3.1.2
flask-cors==6.0.2
geopy==2.4.1
python-dotenv==1.2.1
gunicorn==25.1.0
